create private copy of isl/int.h
We will shortly be introducing a new isl_val abstraction
that is meant to replace isl_int as well as some abuses of isl_qpolynomial.
To implement this new abstraction, we will need some additional
isl_int_* functions.  However, we do not want to publicly expose
these isl_int_* functions since we do not want to encourage their
use by external users.
We therefore create a private copy of isl/int.h where we can
add the additional functions.  Once the transition is complete,
the public version can be removed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>