remove possible use of piplib completely
When the piplib submodule was removed in ff1ea6b (properly remove piplib
submodule, Wed Jan 13 09:59:16 2010 +0100), we still allowed for the use
of an external piplib.  However, the code that supports this use of piplib
has probably not been tested recently and can only hinder further refactorings.
We therefore remove it completely.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>