isl_map_gist_basic_map: remove alignment of context divs with all basic maps
Ever since isl_map_gist was introduced in 0b4619a (add isl_set_gist,
Thu Sep 25 15:24:24 2008 +0200), isl_map_gist (now isl_map_gist_basic_map)
aligns the divs of the context with those of every basic map before
computing the gist of each basic map individually.
Presumably the reason for this global alignment is that if the same
divs appear is multiple basic maps, then the context divs only have
to be aligned to them once instead of once per basic map.
The downside, however, is that if the basic maps have different divs,
then they may end up spreading from one basic map to another,
possibly resulting in a more complicated output, which is contrary
to what you would expect from a "gist" operation.

We remove the global alignment of divs.  If this turns out to be problematic
in some way, we can consider different strategies.

Note that the new test case currently does not trigger even without
this change due to simplifications during parsing of the input.
We keep the test case to illustrate the change and in case parsing changes.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>