isl_{in,}equality_from_aff: skip construction of isl_basic_set
Back when isl_equality_from_aff was introduced in aa6c066
(add isl_equality_from_aff, Thu May 12 15:55:15 2011 +0200),
an isl_constraint was still implemented on top of an isl_basic_map
and we had to construct this isl_basic_set.
Now that isl_constraint is implemented on top of an isl_local_space
and an isl_vec, the construction of this isl_basic_set is pure overhead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>