isl_band_tile: use isl_union_pw_multi_aff_scale_multi_val
We want to remove isl_union_pw_multi_aff_scale_vec, so we have
to use isl_union_pw_multi_aff_scale_multi_val instead.
This means we have to convert the input isl_vec to an isl_multi_val.
We probably want to change the argument of isl_band_tile from
an isl_vec to an isl_multi_val, but this would break compatibility.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>