isl_printer_print_map: make printing order of disjuncts platform independent
In particular, the printing order is based on a sort of the affine hulls
of the disjuncts.  If some of these affine hulls are the same, then the
comparison routine will leave their order open.  Different implementations
of qsort may therefore produce different results.  By using isl_sort instead,
we can avoid this platform dependence.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>