try and avoid depending on mpz_gcdext internals
The results of mpz_gcdext are not uniquely defined and may depend
on the version of GMP used.  Since the results may influence the
construction of the stride offset in detect_stride in isl_ast_build.c,
this may result in user visible differences.

We therefore try to obtain more consistent results by analyzing
the results of the installed version of GMP and calling a modified
version of mpz_gcdext if needed.  This modified version may not
be complete and may still have to be extended.  In the worst case,
we may have to revert to the open-coded version that was removed in
10a7f8a (isl_ast_build.c: use isl_int_gcdext instead of open-coded version,
Wed Apr 17 17:51:55 2013 +0200).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>