scheduler: also (optionally) bound coefficients computed by carry_dependences
2b54bb6 (scheduler: allow to bound the coefficients in the calculated schedule,
Sun Feb 19 14:05:50 2012 +0100) introduced an option to bound the coefficients
of the calculated schedule, but it would only bound the coefficients
of that part of the schedule computed by the core scheduler in
compute_schedule_wcc.

Also make the option apply to that part of the schedule
(if any) computed by the fallback carry_dependences.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>