isl_printer_print_multi_pw_aff: change output format to not lose information
In particular, the original implementation would print the isl_multi_pw_aff
as a list of piecewise affine expressions, but this would lose the information
of the set or output space of the isl_multi_pw_aff.
We now print isl_multi_pw_affs in a similar way to isl_maps, except
that the conditions on the domains are encoded inside the set/output
tuple since each piecewise affine expression may have a different domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>