isl_ast_expr_from_aff: clear denominator before calling extract_modulos
extract_modulos essentially assumes that the denominator of the affine
expression is one.  In particular, it depends on the caller to divide
the extracted modulo expression by the denominator.
Moreover, in a544f6d (isl_ast_expr_from_*: use isl_val,
Tue May 28 15:18:53 2013 +0200), call to isl_aff_get_coefficient
(which extracts the numerator of the coefficient) was replaced
by a call to isl_aff_get_coefficient_val (which extracts the entire,
possibly rational, coefficient).

Rather than trying to extract the numerator from this coefficient,
which is tricky because it may have been reduced already, we impose
that the denominator is one.  The same commit already imposed that
condition on (most of) the body of isl_ast_expr_from_aff.  We now
simply do so before the call to extract_modulos.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>