isl_ast_codegen.c: generate_component: handle non-obviously fixed values
When all domains have an obviously fixed value, we skip the check
for opportunties to expose strides by shifting.  However, the domains
may still have non-obviously fixed values, in which case the gcd
of the strides is zero, which breaks generate_shift_component.
Skip the shifting in this case as well.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>