isl_ast_build_compute_gist_pw_aff: handle parametric domains
When the build domain is a parametric set, the argument to
isl_ast_build_compute_gist_pw_aff does not have a domain and can
therefore not be pulled back.  There is also no need to perform the
pullback since there can be no fixed values for the domain dimensions
if there are no domain dimensions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>