privately export isl_basic_map_reset
This function will be used in the next two commits.
The uses will appear in the same file as the one in which
isl_basic_map_reset is defined, so an alternative would be
to move the function up in the file, but since
isl_map_reset is exported already, it seems to make sense
to export isl_basic_map_reset as well.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>