isl_schedule: explicitly keep track of isl_ctx
When we switch to a schedule tree representation, an isl_schedule
will include an isl_schedule_tree object with an isl_ctx.
For consistency, we will also initialize this embedded isl_ctx
when using the band forest representation.
Start by introducing an isl_ctx directly in isl_schedule so that
we only need to change the location of the isl_ctx when introducing
the schedule tree representation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>