isl_union_set_compute_schedule: fix check for progress in carry_dependences
Commit 6b21b75 (isl_union_set_compute_schedule: ensure carry_dependences
makes progress, Sun Nov 4 16:30:54 2012 +0100) introduced a test
for trivial solutions in order to avoid recomputing the same trivial
solution over and over, but it failed to take into account that
the coefficients computed in carry_dependences correspond to
the original iterators and not on the result of a change of basis.
As a result the test could mistakenly flag some perfectly fine solutions
as trivial and vice versa.
Fix the test by performing the change of basis on the computed solution.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>