basic_set_maximal_difference_at: use isl_basic_map_range_product
The main purpose of this change is to simplify the code.
In particular, the call to the undocumented function
isl_basic_map_from_basic_set is removed by this change.
As a side effect, the result of basic_set_maximal_difference_at may also
be a bit more accurate because isl_basic_map_range_product performs
an isl_basic_map_simplify internally, which the original code did not do.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>