isl_space_underlying: return set space
The function isl_space_underlying is used to treat objects
as if they are not living in a particular space.
In the long term, it is probably better to actually drop
the space, but this is tricky with the current internal representation.
For now, these objects are treated as if they were sets,
so make sure that they actually live in a set space.

In particular, ever since isl-0.07-142-g9043b59e4d (explicitly
differentiate between spaces of maps, sets and parameter sets,
Thu Aug 25 19:21:05 2011 +0200), set spaces are marked explicitly.
Take this marking into account.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>