isl_union_*_mul_isl_int: pass address of isl_int to callback
isl_hash_table_foreach expects a user pointer and our callback function
also dereferences the user pointer. Consequently, we pass in a pointer
to the integer value, not its value.

Note that this commit does not have any effect in practice
because of the way isl_int (i.e., mpz_t) is defined.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>