isl_ast_build: keep track of mapping from internal to input schedule space
We will need this to support the isolate schedule tree AST generation option
which refers to the schedule space of the tree, while, internally,
we may have adjusted the schedule space.

We mainly need to keep track of this mapping in case we decide to scale
down a schedule dimension, but it is also convenient for mapping the
isolate option to the right space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>