add isl_ast_build_node_from_schedule
This allows the user to generate an AST from a schedule specified
as a schedule tree.  The AST generation options in the build are
ignored when building an AST from a schedule tree.
We will be introducing a mechanism for specifying AST generation
options in the schedule tree itself next.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>