add isl_ast_build_alloc
This allows the user to create an isl_ast_build without first having
to create a context set.  This will be especially useful when we will
allow AST generation from a schedule tree with context nodes.
Since the context nodes are embedded inside the schedule tree,
there will typically be no need to specify an additional initial context.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>