isl_ast_build_node_from_schedule: handle basic AST build options
That is, handle the atomic/separate/unroll choice for each schedule
dimension in a band node.  These options are much simpler than their
schedule map representation counterparts.  When using a schedule map
representation, the options had to describe which part of the schedule
tree encoded in the schedule map was affected.  Since the new options
are attached directly to a band node, they only need to describe
which of the schedule dimensions within the band is/are affected.

In theory, the new options are less expressive, but in practice,
the extra expressiveness is not used or even usable.
In particular, the new options are much easier to use by the user
and much easier to handle by isl with a seriously reduced risk
for inconsistencies.

We provide two interfaces for setting these options,
one through a function call and one through an options string.
The first is easier to use programmatically, while the second
is easier to use when loading a schedule tree from a file.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>