make isl_stream->col consistently point to column of next character
The original code did not assign a consistent value to "col",
resulting in the first character on the first line to be considered to be
in column "1", while the first character on other lines would be considered
to be in column "0".  This could lead to confusion, especially
for parsing that is indentation sensitive such as the YAML parsing
that will be introduced next.

Instead of applying a minimal fix, we clean up the code a little
bit to make it more transparent.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>