add isl_constraint_cmp_last_non_zero
This comparison function is somewhat tailored to its use by the AST
generator in the next commit, but it is easier to implement it in
the isl core than in the AST generator itself.  Since we only want
to use public functions in the AST generator, we need to make this
comparison function public too.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>