isl_ast_build_expr_from_basic_set: sort constraints
This should make the AST generation output slightly less sensitive
to the internal representation of core isl.
Moreover, by placing constraints that do not involve integer division
before those that do, we can more easily exploit the earlier constraints
to try and simplify the later constraints in the next commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>