add isl_ast_op_zdiv_r
We currently only extract modulo expressions of type isl_ast_op_pdiv_r
where the argument is known to be non-negative because these can
safely be translated into a C %-operator.  There are however also
some other cases where a mathematical module operation can be replaced
by the C %-operator.  In particular, if we are only interested
in the remainder being zero or not, then the rounding that C performs
does not matter.

We introduce a new isl_ast_op_zdiv_r operation type for such cases.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>