isl_hash_table_find: don't use user specified comparison function to grow table
When isl_hash_table_find needs to grow the hash table, it would
reuse the user specified comparison function to compare entries
against each other.  However, this may lead to undefined behavior
as the comparison function is only meant to be used to compare
entries to the "val" argument.  Moreover, there is no need to
use any comparison function while growing the table since all
entries in the original table are supposed to be distinct already.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>