isl_map_preimage_multi_pw_aff: special case universe set
If the input map does not involve any dimensions in the space that we are
computing the preimage of and if at least one of the isl_pw_aff
in the isl_multi_pw_aff involves more than one piece, then the default
implementation needlessly splits the map along those pieces.
Avoid this splitting by checking for this special case.
The check could be refined by checking individual dimensions
where the corresponding isl_pw_aff involves more than one piece.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>