isl_basic_map_simplify: fix elimination of unit coefficient known divs
If a known div appears with a unit coefficient in a constraint,
then it is eliminated from that constraint in order to simplify
the constraint.
However, one if the div constraints may have been eliminated
as being redundant with respect to the modified constraint and
the elimination of the div from the constraint may therefore
result in a loss of information about the div from the constraints.
To make sure this problem does not occur, we (re)introduce
the div constraint that could potentially have been eliminated
with respect to the modified constraint.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>