isl_union_access_info_compute_flow: determine relative order from schedule tree
That is, if the schedule is given as a schedule tree, then use it directly
to determine the relative order of accesses rather than first converting
the schedule to a union map.  This allows us to exploit the ordering
information explicitly available in the schedule tree rather than having
to rediscover some of it in the extracted union map.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>