optionally use imath for arbitrary precision integers
Some projects such as LLVM and Polly prefer not to use any
LGPL licensed libraries such as GMP.  The optional use of
the MIT licensed IMath library allows them to avoid this.

IMath is included as a submodule to ensure that a single
version of IMath is used for any given version of isl.
The public symbols are prefixed with isl_ to avoid conflicts
with other copies of IMath that may independently be used by
the users of isl.

The deprecated isl_int functions are not supported when IMath is used
since they depend on GMP.  We rename them when IMath is used to ensure
that any attempt to use them in an IMath enabled isl results in
a linker error.

The subdir-objects automake option is turned on since this is the
first time we use source files in subdirectories and since this
option will become the default in future versions of automake.

Signed-off-by: David Peixotto <dpeixott@codeaurora.org>
Signed-off-by: Sebastian Pop <spop@codeaurora.org>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>