isl_map_gist: return universe on gist with empty set
This essentially reverts cc6dfaa (isl_map_gist: return empty set if context
is obviously empty, Fri Oct 28 14:10:46 2011 +0200), where an empty
result was motivated by consistency with the case where the intersection
of input and context is empty.  It is not clear why such a consistency
would be more important than one with the generic case where the context
is a subset of the input set and we return the universe.
Perhaps more importantly, returning a universe makes the gist operation
idempotent in this case.  We do not guarantee that the gist operation
is idempotent in general, but in the current state, if we apply
a gist on a universe with respect to an empty set, we return an empty
set, while if we then compute the gist again with respect to an empty
set, we get a universe again (because now input set and context are
equal), which is surprising behavior to say the least.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>