isl_map_normalize: remove empty parts before sorting
If there are any empty parts in the map, then isl_map_remove_empty_parts
may change the order of the basic maps.  It will then clear the
ISL_MAP_NORMALIZED, so technically the result is not wrong,
but users of isl_map_normalize probably expect the result to
effectively be normalized.
Remove the empty parts first so that the result of isl_map_normalize
is always normalized.

This change should not have any effect on the correctness.
It is not even clear if it will have any effect at all as it
is the result of code review rather than any particular issue.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>