replace CLooG test cases with schedule tree versions
The schedule tree inputs have been recreated from the original
CLooG test cases.

The difference in output for the dealII test case is due to
the fact that the basic sets in the schedule domain are
handled in a different order in separate_schedule_domains.
The different order in turn relates to the different space
that is used for the internal representation of the schedule domain.

Note that in general it is impossible to create two versions
of the input that are guaranteed to produce exactly the same output.
Maintaining two expected outputs for each CLooG test case is not
worth the effort.
The isl_ast_build_node_from_schedule_map interface is still covered
by other test cases.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>