isl_printer_print_pw_aff: keep track of whether we have printed any constraint
In the original code, it was the responsibility of print_basic_set_c to
keep track of whether it has printed any constraints with print_constraint_c
(or otherwise).
However, in d593762 (isl_printer_print_pw_aff: skip constraints implied
by div expression, Tue Feb 7 12:17:03 2012 +0100), we started skipping
some constraints in print_constraint_c.  If this happened to be the
first constraint that we tried to print, then print_basic_set_c's notion
of whether we have printed anything yet would be wrong, resulting in
output that starts with "&&".
Keep track of whether we have printed any constraint in print_constraint_c
itself.

It is difficult to construct a simple test case that reliably tickles
this bug.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>