isl_{set,map}_{,unshifted_}simple_hull: finalize result
These are public function, so the result should be finalized.

The effect on the AST generation output is due to the fact that
during a gist, a simple hull would be computed of the context
that contains a redundant div because the other constraints
involving the div are not retained in the simple hull.
This redundant div would then get introduced in one of the basic
maps that are gisted, where it happens to have a fixed value.
After variable compression, one of the div constraints of the
redundant div would result in an _extra_ constraint in the result.
Now that redundant divs are removed from the simple hull of the context,
this extra constraint is no longer introduced.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>