isl_ast_codegen.c: generate_domain: compute explicit representations of divs
The AST generator cannot build an expression from a set if it has
any unknown divs.  Call isl_set_compute_divs on the guard to
ensure that all divs are known.

This problem only rarely shows up because explicit representations
are usually computed implicitly by the call to isl_ast_build_compute_gist.
However, if the build domain is a universe set (which may happen
in the case of full unrolling), then this implicit divs computation
is not performed.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>