isl_union_map_is_single_valued: look at domain spaces individually
If the input consists of more than one map but is obviously
single-valued on some domain spaces, then we can now detect
the fact it is obviously single-valued on those domain spaces
and avoid potentially costly computations.

Similarly, if the input is not single-valued on at least one domain
space, then there is no need to consider any other domain spaces.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>