isl_space_tuple_match: also ignore parameters in nested spaces
The results of isl_space_tuple_match have been inconsistent.
Due to a call to isl_space_is_equal on nested spaces, the result
would depend on the parameters in case there are any nested spaces,
while it would not depend on the parameters if there are no nested spaces.

In 171cbde (isl_map_is_disjoint: maps with non-matching tuples may still
intersect, Sun Oct 20 10:09:12 2013 +0200), we fixed a caller
of isl_space_tuple_match to not assume that the result does not
depend on the parameters, but indicated that isl_space_tuple_match
may have to be reconsidered.

After a careful investigation of all isl_space_tuple_match calls,
it appears that all callers either make sure the inputs have the same
parameters or assume that the result will not depend on the
parameters.  In particular, isl_ast_build_from_multi_pw_aff
makes this assumption.

Given that the result does not depend on the parameters if
there are no nested spaces, it makes sense to not depend on
them also if there are any nested spaces.

Reported-by: Michael Kruse <meinersbur@googlemail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>