document isl_space_tuple_match as isl_space_tuple_is_equal
The function isl_space_tuple_match is used by the AST generator,
so it should be part of the public API.  Rename it to
isl_space_tuple_is_equal, because "match" is not a very clear
description.

The old name is kept for now in case it was already being used.
It will be removed at some point.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>