isl_union_*_add_*: avoid duplicate checks in callers
isl_union_*_add_* already checks if the new part is "zero", so there
is no need to also perform this check in the caller.

Note that "IS_ZERO" simply checks if the number of pieces
in the piecewise expression is zero, so this can never fail.
We check for failures anyway in case this would ever change
in the future and because the callers used to do this.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>