isl_union_pw_multi_aff_add: compute sum on shared definition domain
The original behavior of isl_union_pw_multi_aff_add was inconsistent.
It would compute the sum on the union of the spaces on which
the arguments are defined, but without one space, it would
compute the sum on the shared definition domain only.

Consistently make isl_union_pw_multi_aff_add compute the sum
on the shared definition domain and add an
isl_union_pw_multi_aff_union_add to (consistently) compute
the sum on the union of the definition domains.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>