split scheduler code from schedule object code
Although the scheduler still has some intimate knowledge
about the current isl_schedule representation, this will
be much less the case after we switch to a schedule tree
representation.  Furthermore, the isl_schedule object will
also become useful to represent schedules not produced by
the scheduler, so it does not really belong in a file that
mainly contains the scheduler code.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>