isl_ast_build_compute_gist: handle parametric domains
When the build domain is a parametric set, the argument to
isl_ast_build_compute_gist does not have a domain and can
therefore not undergo a preimage.  There is also no need to perform the
preimage since there can be no fixed values for the domain dimensions
if there are no domain dimensions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>