add isl_schedule_node_extension
Extension nodes cause the AST generator to generate an AST for
additional domain elements each time the extension node is visited.
They are useful for inserting additional code such as synchronization or
data movement whithin the loops generated from the outer schedule nodes.

This commit only introduces the new node type.
The functions for inserting an extension node in a tree
will be added in subsequent commits.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>