isl_ast_build_set_loop_bounds: do not add stride constraint in eliminated case
We used to add the stride constraint unconditionally, but this meant
that we had to eliminate the current dimension from the build domain
if it turns out that the dimension has a fixed affine value.
Since we have this fixed value, there is no point in adding the
stride constraint since that information will never be used,
especially since we then eliminate the dimension from the stride constraint.
The result of this elimination could however get used and we had
to take special measures to ensure that this information is actually
enforced by the AST.

It is much simpler to not add the stride constraint to the build domain
in the first place.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>