isl_basic_map_gist: remove redundant div constraints
Although this may not be a strict requirement, the output of a gist
operation typically does not contain any redundant constraints
(with respect to other constraints in the output).
If div constraints have been added and subsequently reduced by
equalities involving the corresponding variable, then these
constraints may occasionally be redundant with respect to the
other constraints.  Remove them.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>