isl_ast_codegen.c: create_node_scaled: hoist inclusion of stride guard
In a future commit, we will want to exploit the fact that the stride
guard is going to be added to simplify the guards of the child nodes.
We first need to move the inclusion of the stride guard to a place
where will be able to exploit it later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>