AST generator: exploit all information while looking for unroll lower bound
In particular, the context may include information that is not
completely taken into account in the executed relations because
it may induce extra disjuncts.  Since we are taking the simple
hull for each unrolled iteration anyway, it does not hurt to
include all available information while looking for a lower bound
that can be used for unrolling.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>