isl_ast_codegen.c: create_node_scaled: postpone adding degenerate guards
In the degenerate case, a loop may be printed as an assignment of
the fixed value to the loop iterator.  This assignment does not imply
that the lower bound is actually smaller than or equal to the upper bound.
We therefore need to add this constraint to the graft.

These guards are currently added by add_degenerate_guard from
refine_degenerate, but we will be removing add_degenerate_guard
in a subsequent commit and we need to make sure that these guard
will still be generated.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>