isl_convex_hull.c: convex_hull_pair: add missing return
The problem was introduced all the way back in 7f46696 (isl_convex_hull.c:
convex_hull_pair: skip direction change if result is bounded,
Mon Apr 19 09:23:55 2010 +0200).  Apparently, the code hardly ever
gets executed, which is not entirely unexpected as it is only executed
when the first two basic sets considered in an unbounded set are both
bounded.

Reported-by: antoine.morvan.pro@gmail.com
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>