add ast_always_print_block option
Deep loop nests are often easier to read by humans when all for and if
bodies are printed as blocks (that is, are surrounded by braces in C).
Processing of generated code by e.g. GPUVerify also becomes easier if
braces are emitted uniformly.

The new option provides a way to always print bodies as blocks.

Signed-off-by: Sven van Haastregt <sven.vanhaastregt@arm.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>