isl_coalesce.c: coalesce_subset: properly indicate whether anything has changed
In particular, if the second basic map was dropped, then it was replaced
by another basic map which hasn't been reconsidered yet.  The caller needs
to be notified of this possibility so that it can recheck the basic map
that replaced the one that got dropped.
This appears to have been a typo in a73f222 (isl_map_coalesce: only coalesce
pairs of basic maps with the same divs, Tue Apr 17 19:05:00 2012 +0200),
when coalesce_subset was first introduced.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>