isl_map_coalesce: handle divs that have been simplified away in one basic map
If one of the basic maps satisfies more equalities than the other,
then some of the integer divisions that appear in the other basic map
may have been simplified away in the first.
Try to simplify the extra integer divisions in the other basic map
using the extra equalities in the first to see if we can eliminate
them all.  If so, add those integer divsions with the value specified
by the extra equalities to the first basic map and check if we can
coalesce the result with the other basic map.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>