isl_map_coalesce: coalesce within affine hull first
In principle, any two disjuncts that live in adjacent hyperplanes
can be coalesced together.  First combining basic maps
inside a hyperplane therefore does not affect the potential
for coalescing across the hyperplanes. On the other hand,
first coalescing across hyperplanes results in a basic map
that no longer lives entirely within one of the hyperplanes.
Although it may still be possible to combine this basic map
with other basic maps inside the two hyperplanes, this is no
longer guaranteed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>