isl_map_convex_hull: detect equalities before checking number of basic maps
While detecting equalities, we may discover that some (or even all)
basic maps are empty, so we cannot assume that the number of basic
maps remains the same.  Since the body of isl_map_convex_hull assumes
that there is at least one basic map, we need to detect equalities
before checking the number of basic maps.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>