rename isl_ast_build_domain_type to isl_ast_loop_type
When we introduce basic AST build options in schedule trees,
we will want to be able to set the loop types of individual
schedule dimensions directly and so we will have to expose
this type to the user.  The isl_ast_loop_type name seems
more appropriate, especially considering the names
of the elements in the enumeration.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>