add isl_schedule_node_gist
We will need this function to implement isl_schedule_intersect_domain
since the intersection may make some nodes in the tree unreachable.
We should therefore remove these redundant nodes instead of keeping
track of them.

We currently do not publicly expose this function, but it could
be considered if it turned out to be useful.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>