add isl_union_map_eq_at_multi_union_pw_aff
In principle, it would be possible to construct a map expressing
the order in the target space, map both domain and range of this map
back to the spaces of the union map and perform the intersection
at the end.  However, the intermediate union map could contain
many more maps than are needed.  Only determining the order
for maps that are actually present in the input union map should
be more efficient.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>