isl_coalesce.c: wrap_in_facets: check for number of constraints not increasing
If one of the basic maps sticks out of several constraints of the other
basic maps by at most one, then the total number of wrapping constraints
may be as large

    (number of cut constraints) * (number of constraints of basic map)

Usually, the total number of constraints does not increase since
many of this wrapped constraints will be identical to other wrapped
constraints or valid constraints or they may be redundant with respect
to these constraints.  In fact, in practice, I haven't seen any
case where the number of constraints does increase.

Still, to make sure we will never run into a situation where
the number of constraints increases dramatically, we reject
any fusion that results in an increase in the number of constraints.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>