isl_coalesce.c: can_wrap_in_facet: update comment
When can_wrap_in_facet was first introduced in c4dc489 (isl_map_coalesce:
handle more cases, Mon Feb 8 19:24:16 2010 +0100), it was called
from two contexts.  One of these calls was removed in e4ac1a5
(isl_map_coalesce: allow wrapping in sets that stick out in different
directions, Sun Apr 18 17:50:05 2010 +0200), but the documentation
was not udpated accordingly.
As the updated comment states, we no longer need to wrap
around the facet of the basic map with the inequality adjacent
to the other basic map, but we keep the code because we
will be able to reuse it later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>