isl_printer_flush: actually clear buffer when called on string printer
isl_printer_flush is documented as clearing the buffer when called
on a string printer, but ever since it was introduced in 6c91cdb
(add isl_printer_flush, Sat Apr 24 17:43:32 2010 +0200), it would
only reset the number of elements in the buffer, leaving it to the next
write to actually clear the buffer.
Clear the buffer immediately.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>