isl_ast_codegen.c: add_implied_guards: combine degenerate and stride constraints
If there are both stride constraints and degenerate loop bounds, then
a combination of them may imply more guards then the two sets of
constraints would imply separately.  Combine them before deriving
the implied guards.

It is difficult to find a test case that illustrates this issue
in the current code base, but it has occurred in an experimental
schedule tree branch.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>