isl_coalesce.c: check_eq_adj_eq: allow multiple equalities adjacent to equality
The limitation to one equality adjacent to an equality was introduced
in 59baf90 (isl_map_coalesce: handle some cases of pairs of adjacent
equalities, Wed Oct 13 22:02:25 2010 +0200) to avoid introducing
large coefficients when coalescing some cases of multiple equalities
adjacent to an equality.  In 239513e (isl_map_coalesce: optionally bound
the coefficients of wrapping constraints, Wed Apr 11 10:54:37 2012 +0200),
a generic mechanism was introduced for (optionally) avoidind such large
coefficients, so we no longer need this restriction.

If there is more than one such equality, then the coalesced result
will satisfy one or more equalities that are not explicitly available
in the input.  Instead, they will appear as pairs of inequalities
in the wrapped constraints and need to be changed into explicit
equalities.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>