isl_coalesce.c: fuse: copy valid half of equality
If an equality as a whole is not valid for the other basic map,
but one of the two inequalities that make up the equality is valid,
then copy the valid part.  This ensures that all valid inequalities
are added to the fused basic map such that we can avoid wrapping
valid inequalities in a future commit without having the inequality
disappear.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>