isl_coalesce.c: coalesce_with_subs: add div constraints to tableau
Since 81fed9f (isl_map_coalesce: handle divs that have been simplified away
in one basic map, Fri Oct 10 11:49:16 2014 +0200), we try adding
back divs to a basic map from another basic map that may have been
simplified away using the equalities in the basic map.
When we add divs from one basic map to the other, we also add
the corresponding div constraints.  These need to be added
to the tableau as well since ineq_status_in first checks
if the given constraint is redundant in its own basic map
by checking if the corresponding constraint in the tableau
is redundant.  This assumes that there is such a constraint
in the tableau and that it corresponds to the constraint
in the basic map.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>