Add public access to short-circuit && and ||
This allows the user to build ast expressions, e.g., runtime checks,
that should only be conditionally executed. This is especially important
if the conditional part of the expression is not pure but also if it
might be costly to evaluate the expression unconditionally.

In Polly these functions are used to build runtime checks for assumed
invariant loads. As the checks might depend on the value initially
stored in the assumed invariant location we have to perform a load of
that location. However, this is only sound if the original load is
guaranteed to be executed at some point. If not we might even access an
invalid pointer and crash the program. The assumptions that will
guarantee execution of the original load are therefor used as condition
for the actual runtime check.

Signed-off-by: Johannes Doerfert <doerfert@cs.uni-saarland.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>