isl_schedule_constraints_compute_schedule: handle conditional validity after band
Conditional validity constraints are currently only handled during
the construction of a band.  That is, only within the construction
of a band does the scheduler ensure that there are no non-local
conditions that are adjacent to violated conditional validity constraints.
However, a band may still result in non-local conditions and then
the conditional validity constraints need to be enforced by inner bands.
This issue does not show up very frequently because by the time
the conditions become non-local, the adjacent conditional validity
constraints have usually already been removed from the dependence graph,
but it does still happen.

A test case will be added on the master branch.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>