isl_scheduler.c: compute_schedule_wcc: extract out compute_schedule_wcc_partial
This further reduces the size of compute_schedule_wcc and, together with
the previous commit, this allows us to perform additional computations
between the pure construction of a band and the incoporation of
that band in the schedule.  This will allow us in particular
to attempt to combine strongly connected components first.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>