isl_schedule_constraints_compute_schedule: allow incremental combination of SCCs
Unless the schedule_serialize_sccs option is set, an entire (weakly)
connected component is currently scheduled as a whole.
This may result in schedule problems with a large number of variables,
which in turn may result in high scheduling times, and it makes it
more difficult to control how strongly connected components
in the dependence graph should be combined as everything needs
to be encoded in a single (I)LP problem.

Introduce an alternative way of computing a schedule for a weakly
connected component where a partial schedule is first computed
for each strongly connected component (SCC) within the weakly connected
component separately and where these SCCs are subsequently combined
into clusters by scheduling the SCCs (or clusters of SCCs) with respect
to each other.  This usually results in a smaller number of variables
since the number of statements in an SCC and the number of clusters
that are being combined together in one go is usually smaller than
the total number of statements in the weakly connected component.
There is also more direct control over which SCCs should be combined
together.  In particular, in the next commit this flexibility
will be exploited to (optionally) only combine clusters if this does
not reduce the level of coincidence.

The heuristic used to select pairs of clusters to be merged
has only received limited investigation.
The current heuristic favors the combination of bands that are
linked by a proximity edge that allows many schedule dimensions
to be aligned and that are not far apart in the topological order of
the strongly connected components.

The merge of two (or more) clusters is only accepted if their
is at least one proximity constraint between two clusters
that is completely optimized (up to a small constant).

An alternative for scheduling clusters with respect to each other
would be to reschedule all statements in the clusters that are
being merged.  This would typically require more variables,
but it would allow for schedules that are now being missed because
the schedule of an SCC is fixed once it has been computed.

A schedule_whole_component option is introduced to choose between
the old and the new algorithm, which is default on, corresponding
to the original behavior.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>